tests: measure coverage
To measure coverage in a subprocess `atexit` handlers must be run. They
are unforutnately skipped if the application is terminated from the Qt
code.

A workaround was added so Python is able to handle signals by using a
dummy timer which gives the interpreter an oppurtunity to do its job.