Execute browser via `multiprocessing`
The browser window runs its separate process in order to not let `PyQt`
pollute the root process and make the core able to still use `asyncio`
without any hassle.

Unfortunately spawning a separate Python instance is not so trivial as
it won't inherit all state from the parent process. It makes the
application harder to integrate in more exotic deployments such as in `NixOS`