Run openconnect outside the event loop
By the time we run openconnect there's no need to use asyncio any more:
The async GUI interactions are complete, and we simply want to run
openconnect as a subprocess and wait for its result. Move this part of
the program (as well as some other "sync" stuff) out of the async
`_run()` and into the sync `run()` function.

This fixes/simplifies the control flow when the users aborts with Ctrl+C
while openconnect is running: Without it, asyncio's run_until_complete()
will abort the asyncio event loop on SIGINT without running any exception
handlers or finally clasues.

This is in preparation for adding code to be run when openconnect is
being aborted.

Suggested-by: László Vaskó <laszlo.vasko@outlook.com>